+++ contexts/AuthContext.tsx
@@ -24,6 +24,10 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     supabase.auth.getSession().then(({ data: { session } }) => {
       setUser(session?.user ?? null);
       setLoading(false);
+    }).catch((error) => {
+      console.error('Error getting session:', error);
+      setUser(null);
+      setLoading(false);
     });
 
     // Listen for auth changes
