+++ lib/auth.ts
@@ -11,6 +11,19 @@ export async function signIn(email: string, password: string) {
   return data;
 }
 
+export async function signInWithGoogle() {
+  const supabase = createClient();
+  const { data, error } = await supabase.auth.signInWithOAuth({
+    provider: 'google',
+    options: {
+      redirectTo: `${window.location.origin}/auth/callback`
+    }
+  });
+  
+  if (error) throw error;
+  return data;
+}
+
 export async function signUp(email: string, password: string, full_name: string) {
   const supabase = createClient();
   const { data, error } = await supabase.auth.signUp({
